ParticipacaoMissao completo; alteração em Missao

+++ src/main/java/com/example/guilda/dto/aventura/MissaoRelatorioDTO.java
@@ -3,11 +3,13 @@ package com.example.guilda.dto.aventura;
 import com.example.guilda.domain.aventura.StatusMissao;
 import com.example.guilda.domain.aventura.NivelPerigo;
 
+import java.math.BigDecimal;
+
 public record MissaoRelatorioDTO(
         Long id,
         String titulo,
         StatusMissao status,
         NivelPerigo nivelPerigo,
         Long totalParticipantes,
-        Long totalRecompensa
+        BigDecimal totalRecompensa
 ) {}
